+++ pcsx2/GS/Renderers/SW/GSVertexSW.h
@@ -246,7 +246,7 @@ struct alignas(32) GSVertexSW
 #endif
 	}
 
-	typedef void (*ConvertVertexBufferPtr)(const GSDrawingContext* RESTRICT ctx, GSVertexSW* RESTRICT dst, const GSVertex* RESTRICT src, size_t count);
+	typedef void (*ConvertVertexBufferPtr)(const GSDrawingContext* RESTRICT ctx, GSVertexSW* RESTRICT dst, const GSVertex* RESTRICT src, u32 count);
 
 	static ConvertVertexBufferPtr s_cvb[4][2][2][2];
